+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project attempts to match the major and minor versions of [stactools](https://github.com/stac-utils/stactools) and increments the patch number as needed.
 
-## [Unreleased] - TBD
+## [0.4.0] - 2023-04-14
 
 ## Added
 
@@ -46,7 +46,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - remove python 3.7 support
 
-[Unreleased]: https://github.com/stactools-packages/cop-dem/compare/v0.3.0..main
+[Unreleased]: https://github.com/stactools-packages/cop-dem/compare/v0.4.0..main
+[0.4.0]: https://github.com/stactools-packages/cop-dem/compare/v0.3.0..v0.4.0
 [0.3.0]: https://github.com/stactools-packages/cop-dem/compare/v0.2.0..v0.3.0
 [0.2.0]: https://github.com/stactools-packages/cop-dem/compare/v0.1.1..v0.2.0
 [0.1.1]: https://github.com/stactools-packages/cop-dem/compare/v0.1.0..v0.1.1